requirements.txt
# XAUUSD Signal Engine Dependencies
# Python 3.11 recommended

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # For parquet file support

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Utilities
joblib>=1.3.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Live trading dependencies
websocket-client>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0
